scripts/egress_alert_fn/requirements.txt
functions-framework==3.*
google-cloud-monitoring>=2.20
cloudevents
